Implement beginning-end ignore optimization
Improve efficiency by reducing problem set and therefore the c matrix.
This helps to significantly reduce cpu time and memory consumption for comparing big lists with common items at the beginning and the end of them.
Includes test to check and debug such cases. Without the optimization it fails due to OutOfMemoryException.